Add files via upload
Added an ifdef to separate between linux and windows. Linux code uses the flexiblas routine "mkl_set_num_threads" (which isn't on Windows) to set the number of threads to 1 (using more seemed to slow dgeesv down rather than to speed it up!). Windows does not seem to have this problem.