Debug on loglik calculations for RW->OU
Removed restriction on calculation of W=inv(Lambda)exp(Lambda*t)*V*m (which is used for calculating the stepsize expected increment, u_t) on having pull in the last layer, as eigenvector decompoosition scrambles the order of rows/columns anyway. This restriction made for a jump in log-likelihood calculations for a model with RW causally connected to an OU-like tracking, when the connection factor went from 0 to non-zero. Also added a test in layer.series.structure for whether the user has sepcified no pull (i.e. RW at the bottom layer) without setting an initial state treatment. If the warning is necessary, the initial treatment is then automatically set.