Add files via upload
Removed restriction on calculation of W=inv(Lambda)exp(Lambda*t)*V*m on having pull in the last layer, as eigenvector decompoosition scrambles the order of rows/columns anyway. This restriction made for a jump in log-likelihood calculations for a model with RW causally connected to an OU-like tracking, when the connection factor went from 0 to non-zero.