Add files via upload
Debug on treatment of combination of causal models and processes with and without linear dependency. (Calculation of W vectors in initial state and for each step. Removed test for linear dependency within the calculations, as that depended on the eigenvalue decomposition. Put it into a separate vector calculated earlier instead.)